refactor: [165] moved some files to root
