--- fenn/args/__init__.py
@@ -1,3 +1,0 @@
-from fenn.args.parser import Parser
-
-__all__ = ["Parser"]
