--- fenn/core/__init__.py
@@ -1,3 +1,0 @@
-from fenn.core.fenn import Fenn
-
-__all__ = ["Fenn"]
