+++ fenn/experimental/vision/image_dir_summary.py
@@ -1,7 +1,7 @@
 from collections import Counter
 from pathlib import Path
 from typing import Any, Dict, List, TypedDict
-from fenn.utils.logging import logger
+from fenn.logging import logger
 
 try:
     from PIL import Image
