--- fenn/export/__init__.py
@@ -1,3 +1,0 @@
-from fenn.export.exporter import Exporter
-
-__all__ = ["Exporter"]
