+++ fenn/nn/trainers/classification_trainer.py
@@ -19,8 +19,8 @@ from sklearn.metrics import (  # noqa: F401
 )
 from torch.utils.data import DataLoader
 
+from fenn.logging import logger
 from fenn.nn.utils import Checkpoint
-from fenn.utils.logging import logger
 
 from .trainer import Trainer
