+++ fenn/nn/trainers/regression_trainer.py
@@ -14,8 +14,8 @@ from rich.progress import (
 from sklearn.metrics import r2_score
 from torch.utils.data import DataLoader
 
+from fenn.logging import logger
 from fenn.nn.utils import Checkpoint
-from fenn.utils.logging import logger
 
 from .trainer import Trainer
