+++ fenn/nn/utils/checkpoint.py
@@ -3,8 +3,8 @@ from typing import List, Optional, Union
 
 import torch
 
+from fenn.logging import logger
 from fenn.nn.utils.state import TrainingState
-from fenn.utils.logging import logger
 
 
 class Checkpoint:
