--- fenn/secrets/__init__.py
@@ -1,3 +1,0 @@
-from .keystore import KeyStore
-
-__all__ = ["KeyStore"]
